requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.1.0.post1
requests==2.31.0
python-dotenv==1.0.0
fastapi-mail==1.4.1
aiofiles==23.2.1
python-magic==0.4.27